internal: remove pre-Go 1.18 workaround for reading build info
We haven't supported Go 1.17 or older since February 2023, when Go 1.20
was released. So we don't need the workaround for reading debug info
from the binary for pre-Go 1.18 builds.
