+++ internal/gitmetadata.go
@@ -8,7 +8,10 @@ package internal
 import (
 	"net/url"
 	"os"
+	"runtime/debug"
 	"sync"
+
+	"gopkg.in/DataDog/dd-trace-go.v1/internal/log"
 )
 
 const (
@@ -71,6 +74,32 @@ func getTagsFromDDTags() map[string]string {
 		TagCommitSha:     etags[TagCommitSha],
 		TagGoPath:        etags[TagGoPath],
 	}
+}
+
+// getTagsFromBinary extracts git metadata from binary metadata
+func getTagsFromBinary() map[string]string {
+	res := make(map[string]string)
+	info, ok := debug.ReadBuildInfo()
+	if !ok {
+		log.Debug("ReadBuildInfo failed, skip source code metadata extracting")
+		return res
+	}
+	goPath := info.Path
+	var vcs, commitSha string
+	for _, s := range info.Settings {
+		if s.Key == "vcs" {
+			vcs = s.Value
+		} else if s.Key == "vcs.revision" {
+			commitSha = s.Value
+		}
+	}
+	if vcs != "git" {
+		log.Debug("Unknown VCS: '%s', skip source code metadata extracting", vcs)
+		return res
+	}
+	res[TagCommitSha] = commitSha
+	res[TagGoPath] = goPath
+	return res
 }
 
 // GetGitMetadataTags returns git metadata tags
